moved EnumSearchName definition
